+++ CHANGES.md
@@ -1,6 +1,12 @@
 CHANGES
 =======
 
+- `Client` class
+    * Added `getAttributes()` method.
+    * Added `setAttributes(Pair[])` method.
+    * Added `setAttributes(Iterable<Pair>)` method.
+
+
 2.86 (2021-01-23)
 -----------------